requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-filter>=23.1
psycopg2-binary
drf-yasg>=1.21
django-cors-headers
boto3
django-storages
python-dotenv